V1.0.3

+++ create-next-fullstack/src/index.js
@@ -5,8 +5,25 @@ const chalk = require("chalk");
 const fs = require("fs-extra");
 const path = require("path");
 const { execSync } = require("child_process");
+const prompts = require("prompts");
+
+const setup = [
+  {
+    type: "text",
+    initial: "my-app",
+    name: "projectName",
+    message: "What is the name of your project?",
+  },
+  {
+    type: "confirm",
+    name: "autoInstall",
+    message: "Install packages automatically?",
+    initial: true,
+  },
+];
 
 let projectName;
+let autoInstall;
 let startTime = Date.now();
 
 /*
@@ -15,35 +32,19 @@ let startTime = Date.now();
 
 const program = new commander.Command(packageJson.name)
   .version(packageJson.version)
-  .arguments("<project-directory>")
-  .usage(`${chalk.green("<project-directory>")}`)
-  .action((name) => {
-    projectName = name;
+  .usage(`${chalk.green("Create a Fullstack project")}`)
+  .action(() => {
+    prompts(setup).then((answer) => {
+      if (!answer.projectName) {
+        console.log("exit");
+        process.exit(0);
+      }
+      projectName = answer.projectName;
+      autoInstall = answer.autoInstall;
+      createProject();
+    });
   })
-  .option("--chakra", "use chakra ui")
   .parse(process.argv);
-
-if (typeof projectName === "undefined") {
-  console.error("[Error] Please specify the project directory:");
-  console.log(
-    `  ${chalk.cyan(program.name())} ${chalk.green("<project-directory>")}`
-  );
-  process.exit(1);
-}
-
-const projectDestination = path.join(process.cwd(), projectName);
-
-if (fs.existsSync(projectDestination)) {
-  console.error(
-    `[Error] The directory ${chalk.green(projectName)} already exists.`
-  );
-  process.exit(1);
-}
-
-fs.copySync(
-  path.join(__dirname, "..", program.chakra ? "chakra" : "default"),
-  projectName
-);
 
 const useYarn = () => {
   try {
@@ -54,39 +55,60 @@ const useYarn = () => {
   }
 };
 
-const yarn = useYarn();
+const createProject = () => {
+  const projectDestination = path.join(process.cwd(), projectName);
 
-/*
- *   Install server dependencies & create .gitignore
- */
+  if (fs.existsSync(projectDestination)) {
+    console.error(
+      `[Error] The directory ${chalk.green(projectName)} already exists.`
+    );
+    process.exit(1);
+  }
 
-console.log(chalk.green("Installing server dependencies"));
+  fs.copySync(
+    path.join(__dirname, "..", program.chakra ? "chakra" : "default"),
+    projectName
+  );
 
-process.chdir(projectDestination + "/server");
+  const yarn = useYarn();
 
-fs.writeFileSync(
-  ".gitignore",
-  `node_modules 
+  /*
+   *   Install server dependencies & create .gitignore
+   */
+
+  process.chdir(projectDestination + "/server");
+
+  fs.writeFileSync(
+    ".gitignore",
+    `node_modules 
    dist`
-);
+  );
 
-if (yarn) {
-  execSync("yarn install", { stdio: [0, 1, 2] });
-} else {
-  execSync("npm install", { stdio: [0, 1, 2] });
-}
+  if (autoInstall) {
+    console.log(chalk.green("Installing server dependencies..."));
+    if (yarn) {
+      execSync("yarn install", {
+        //stdio: ["pipe", "pipe", process.stderr]
+        stdio: ["pipe"],
+      });
+    } else {
+      execSync("npm install", {
+        //stdio: ["pipe", "pipe", process.stderr]
+        stdio: ["pipe"],
+      });
+    }
+    console.log(chalk.green("Done..."));
+  }
 
-/*
- *   Install web dependencies & create .gitignore
- */
+  /*
+   *   Install web dependencies & create .gitignore
+   */
 
-console.log(chalk.green("Installing web dependencies"));
+  process.chdir(projectDestination + "/web");
 
-process.chdir(projectDestination + "/web");
-
-fs.writeFileSync(
-  ".gitignore",
-  `# dependencies
+  fs.writeFileSync(
+    ".gitignore",
+    `# dependencies
 /node_modules
   
 #next
@@ -111,28 +133,50 @@ yarn-error.log*
 # typescript
 *.tsbuildinfo
 next-env.d.ts`
-);
+  );
 
-if (yarn) {
-  execSync("yarn install", { stdio: [0, 1, 2] });
-} else {
-  execSync("npm install", { stdio: [0, 1, 2] });
-}
-console.log();
-console.log();
-console.log(
-  chalk.green(`Initialized project in ${(Date.now() - startTime) / 1000}s`) +
-    " 🎉"
-);
-console.log();
-console.log();
-console.log(chalk.underline("Next Steps:"));
-console.log();
-console.log(chalk.italic(`cd ${projectName}/server`));
-console.log(
-  chalk.italic(yarn ? "yarn watch && yarn dev" : "npm run watch && npm run dev")
-);
-console.log();
-console.log(chalk.italic(`cd ${projectName}/web`));
-console.log(chalk.italic(yarn ? "yarn dev" : "npm run dev"));
-console.log();
+  if (autoInstall) {
+    console.log(chalk.green("Installing web dependencies..."));
+    if (yarn) {
+      execSync("yarn install", {
+        //stdio: ["pipe", "pipe", process.stderr],
+        stdio: ["pipe"],
+      });
+    } else {
+      execSync("npm install", {
+        //stdio: ["pipe", "pipe", process.stderr]
+        stdio: ["pipe"],
+      });
+    }
+    console.log(chalk.green("Installing server dependencies..."));
+  }
+
+  console.log();
+  console.log();
+  console.log(
+    chalk.green(`Initialized project in ${(Date.now() - startTime) / 1000}s`) +
+      " 🎉"
+  );
+  console.log();
+  console.log();
+  console.log(chalk.underline("Next Steps:"));
+  console.log();
+  console.log(chalk.dim(`cd ${projectName}/server`));
+  console.log(
+    chalk.dim(
+      yarn
+        ? `${autoInstall ? "" : "yarn install && "}yarn watch && yarn dev`
+        : `${autoInstall ? "" : "npm install && "}npm run watch && npm run dev`
+    )
+  );
+  console.log();
+  console.log(chalk.dim(`cd ${projectName}/web`));
+  console.log(
+    chalk.dim(
+      yarn
+        ? `${autoInstall ? "" : "yarn install && "}yarn dev`
+        : `${autoInstall ? "" : "npm install && "}npm run dev`
+    )
+  );
+  console.log();
+};
